fix a class name in jdbc module and reformat
