+++ jdbc/src/main/java/org/apache/iotdb/jdbc/IoTDBDataSource.java
@@ -28,119 +28,117 @@ import org.slf4j.LoggerFactory;
 
 public class IoTDBDataSource implements DataSource {
 
-    private final Logger logger = LoggerFactory.getLogger(IoTDBDataSource.class);
+  private final Logger logger = LoggerFactory.getLogger(IoTDBDataSource.class);
 
-    private String url;
-    private String user;
-    private String password;
-    private Properties properties;
-    private Integer port = 6667;
+  private String url;
+  private String user;
+  private String password;
+  private Properties properties;
+  private Integer port = 6667;
 
-    public IoTDBDataSource() {
-        properties = new Properties();
+  public IoTDBDataSource() {
+    properties = new Properties();
+  }
+
+  public IoTDBDataSource(String url, String user, String password, Integer port) {
+    this.url = url;
+    this.properties = new Properties();
+    properties.setProperty("user", user);
+    properties.setProperty("password", password);
+    if (port != 0) {
+      this.port = port;
     }
+  }
 
-    public IoTDBDataSource(String url, String user, String password, Integer port) {
-        this.url = url;
-        this.properties = new Properties();
-        properties.setProperty("user",user);
-        properties.setProperty("password",password);
-        if(port!=0) {
-            this.port = port;
-        }
+  public String getUser() {
+    return user;
+  }
+
+  public void setUser(String user) {
+    this.user = user;
+    properties.setProperty("user", user);
+  }
+
+  public String getPassword() {
+    return password;
+  }
+
+  public void setPassword(String password) {
+    this.password = password;
+    properties.setProperty("password", password);
+  }
+
+  public Integer getPort() {
+    return port;
+  }
+
+  public void setPort(Integer port) {
+    this.port = port;
+  }
+
+  public String getUrl() {
+    return url;
+  }
+
+  public void setUrl(String url) {
+    this.url = url;
+  }
+
+  @Override
+  public Connection getConnection() throws SQLException {
+    try {
+      return new IoTDBConnection(url, properties);
+    } catch (TTransportException e) {
+      e.printStackTrace();
     }
+    return null;
+  }
 
-    public String getUser() {
-        return user;
+  @Override
+  public Connection getConnection(String username, String password) throws SQLException {
+    try {
+      Properties newProp = new Properties();
+      newProp.setProperty("user", username);
+      newProp.setProperty("password", password);
+      return new IoTDBConnection(url, newProp);
+    } catch (Exception e) {
+      e.printStackTrace();
     }
+    return null;
+  }
 
-    public void setUser(String user) {
-        this.user = user;
-        properties.setProperty("user",user);
-    }
+  @Override
+  public PrintWriter getLogWriter() throws SQLException {
+    return null;
+  }
 
-    public String getPassword() {
-        return password;
-    }
+  @Override
+  public void setLogWriter(PrintWriter printWriter) throws SQLException {
 
-    public void setPassword(String password) {
-        this.password = password;
-        properties.setProperty("password",password);
-    }
+  }
 
-    public Integer getPort() {
-        return port;
-    }
+  @Override
+  public void setLoginTimeout(int i) throws SQLException {
 
-    public void setPort(Integer port) {
-        this.port = port;
-    }
+  }
 
-    public String getUrl() {
-        return url;
-    }
+  @Override
+  public int getLoginTimeout() throws SQLException {
+    return 0;
+  }
 
-    public void setUrl(String url) {
-        this.url = url;
-    }
+  @Override
+  public java.util.logging.Logger getParentLogger() throws SQLFeatureNotSupportedException {
+    return null;
+  }
 
-    @Override
-    public Connection getConnection() throws SQLException {
-        try {
-            return new IoTDBConnection(url, properties);
-        } catch (TTransportException e) {
-            e.printStackTrace();
-        }
-        return null;
-    }
+  @Override
+  public <T> T unwrap(Class<T> aClass) throws SQLException {
+    return null;
+  }
 
-    @Override
-    public Connection getConnection(String username, String password) throws SQLException {
-       try {
-           Properties newProp = new Properties();
-           newProp.setProperty("user",username);
-           newProp.setProperty("password",password);
-           return new IoTDBConnection(url, newProp);
-       }
-       catch (Exception e){
-           e.printStackTrace();
-       }
-       return null;
-    }
-
-
-    @Override
-    public PrintWriter getLogWriter() throws SQLException {
-        return null;
-    }
-
-    @Override
-    public void setLogWriter(PrintWriter printWriter) throws SQLException {
-
-    }
-
-    @Override
-    public void setLoginTimeout(int i) throws SQLException {
-
-    }
-
-    @Override
-    public int getLoginTimeout() throws SQLException {
-        return 0;
-    }
-
-    @Override
-    public java.util.logging.Logger getParentLogger() throws SQLFeatureNotSupportedException {
-        return null;
-    }
-
-    @Override
-    public <T> T unwrap(Class<T> aClass) throws SQLException {
-        return null;
-    }
-
-    @Override
-    public boolean isWrapperFor(Class<?> aClass) throws SQLException {
-        return false;
-    }
+  @Override
+  public boolean isWrapperFor(Class<?> aClass) throws SQLException {
+    return false;
+  }
 }
